AbstractLogService: refactor level check logic
We do the level check in the level-specific methods now, in preparation
for the protected log(int level, Object msg, Throwable t) method to
become the central override point of concrete implementations. That way,
specific implementations will not need to worry about filtering based on
level in the typical case.